Change hprotp into a suffix encoding
By reversing the parsing direction for decoding hproto messages, the
encoding of those messages becomes substantially easier and more
efficient.

In fact, the message-encoding can now be done in a single pass, and does
not require any copying/moving of the data once they have been written
into the output buffer for the first time.

Also change the available external size widths from a linearly
increasing byte counts to powers of two.

That means, instead of sizes 1, 2, 3 and 4 bytes, not only 1, 2, 4 or 8
bytes are supported as field widths of the external content size fields.