Reverse encoding order
This is a *major* change to the definition, making the new versison
completely binary-incompable with earlier encodings!

However, I think the advantages of the new approach outweigh that
disadvantage. Also, I have not actually used hproto in any application
yet, making the change easy for me.

Also started conversion of the document into asciidoc syntax.