Add UUIDs of externally referenced types
This makes the descriptions more reliable, because the same type names
might map to different UUIDs in different projects.

By adding the UUIDs of all referenced types from other documents to
every description document for a particular type, the associations
become unambigous and globally unique irrespective of the local context.